Defense ready
Decided to rework a bit of the code so that it works a tad bit more consistently. Wasn't done in the last second, just remembered to update it now. Also added a function to list all the available keys through bluetooth! Anyway I don't know if this will be my final update but we'll see after the defense.